+++ include/GafferScene/TypeIds.h
@@ -198,6 +198,7 @@ enum TypeId
 	ReflectionConstraintTypeId = 120153,
 	CurvesInterpolationTypeId = 120154,
 	MeshLightTypeId = 120155,
+	CurvesTangentsTypeId = 120156,
 
 	LastTypeId = 120999
 };